[PATCH] uml: 64-bit cleanups in the system calls
64-bit cleanup - this fixes the return values of the system calls to be longs.

+++ arch/um/include/syscall_user.h
@@ -7,7 +7,7 @@
 #define __SYSCALL_USER_H
 
 extern int record_syscall_start(int syscall);
-extern void record_syscall_end(int index, int result);
+extern void record_syscall_end(int index, long result);
 
 #endif
